WIP feat(decimal): cascade migration entities + ports core (EXP-003)
Couches migrées vers rust_decimal::Decimal (ADR-0007/0008):
- Domain entities: journal_entry, charge_distribution, unit, unit_owner
- Application ports: JournalEntryRepository, ChargeDistributionRepository,
  UnitOwnerRepository (HashMap<String,f64> → HashMap<String,Decimal>,
  ownership_percentage tuple → Decimal, get_total_*_due → Decimal)
- Application services: expense_accounting_service (vat compares,
  vat_rate * 100, tests via dec!())
- Application use_cases: charge_distribution_use_cases (mocks + tests),
  unit_use_cases tests (Decimal quota), etat_date_use_cases helpers
- Infrastructure repositories: journal_entry_repository_impl
  (suppression Decimal::from_f64_retain - line.debit est déjà Decimal),
  charge_distribution_repository_impl (binds Decimal direct)
- DTOs: unit_dto.rs quota → Decimal

Tests added:
- @edge: edge_decimal_exactness_preserved_on_cumul (journal_entry)
  prouve 0.1 + 0.2 + 0.3 = 0.6 exact en Decimal (IEEE 754 fail)
- @edge: edge_distribution_decimal_exactness (charge_distribution)
- @negative: negative_debit_amount_rejected (journal_entry)

Cascade RESTANTE (compile NOK ~50 errors):
- domain/services/expense_calculator (Unit.quota)
- domain/services/annual_report_exporter (sums + add-assign)
- domain/services/owner_statement_exporter (add-assign)
- application/use_cases/dashboard_use_cases (sums)
- infrastructure/web/handlers/{building,expense,mcp_sse}_handlers
- application/use_cases/etat_date_use_cases (total_quota: f64,
  EtatDate entity 7+ champs f64)
- infrastructure/database/repositories/unit_owner_repository_impl
  (SQL ownership_percentage DOUBLE PRECISION → besoin migration NUMERIC)
- BDD/integration tests

Stratégie next steps:
1. Migration SQL: unit_owners.ownership_percentage DOUBLE→NUMERIC(6,5)
2. cargo sqlx prepare pour regenerer offline cache
3. EtatDate entity: 7 champs monétaires f64 → Decimal
4. Handlers + use_cases sums via .sum::<Decimal>()
5. Tests BDD/integration: dec!() literals partout

Anti-pattern explicitement évité: pas de .to_f64()/from_f64_retain
ad-hoc qui contredirait ADR-0007. Migration systémique.

Refs: #436, #433

+++ backend/src/application/ports/charge_distribution_repository.rs
@@ -1,5 +1,6 @@
 use crate::domain::entities::ChargeDistribution;
 use async_trait::async_trait;
+use rust_decimal::Decimal;
 use uuid::Uuid;
 
 #[async_trait]
@@ -30,5 +31,5 @@ pub trait ChargeDistributionRepository: Send + Sync {
     async fn delete_by_expense(&self, expense_id: Uuid) -> Result<(), String>;
 
     /// Get total amount due for an owner across all pending distributions
-    async fn get_total_due_by_owner(&self, owner_id: Uuid) -> Result<f64, String>;
+    async fn get_total_due_by_owner(&self, owner_id: Uuid) -> Result<Decimal, String>;
 }
